refactor: consolidate React Fast Refresh middleware into pluginDevRemoteHmr
Move /@react-refresh proxy middleware from standalone pluginSharedReactRefresh
into pluginDevRemoteHmr's configureServer hook, per maintainer feedback.

The middleware sits in the if (isRemote) block alongside the existing
__mf_hmr metadata endpoint. No behavioral change — same proxy module,
same activation conditions (remote + remoteHmr enabled).

Deleted: pluginSharedReactRefresh.ts and its test file.
Moved: 4 test cases into pluginDevRemoteHmr.test.ts.
