+++ src/plugins/pluginDevRemoteHmr.ts
@@ -7,6 +7,26 @@ const REMOTE_HMR_EVENT = 'mf:remote-update';
 const REMOTE_HMR_CONNECT_RETRY_DELAY_MS = 1000;
 const REMOTE_HMR_CONNECT_MAX_RETRIES = 10;
 
+/**
+ * Proxy module served for `/@react-refresh` on MF remote dev servers.
+ * Delegates to the host page's RefreshRuntime instance via
+ * `window.location.origin`, ensuring a single shared component registry
+ * across federation boundaries. A `configureServer` middleware is used
+ * instead of `resolveId` because `@vitejs/plugin-react`'s
+ * `vite:react-refresh` sub-plugin uses `enforce: 'pre'` and typically
+ * wins the `resolveId` race.
+ */
+const REACT_REFRESH_PROXY_MODULE = [
+  `const __rt = await import(window.location.origin + '/@react-refresh');`,
+  `export const injectIntoGlobalHook = __rt.injectIntoGlobalHook;`,
+  `export const register = __rt.register;`,
+  `export const createSignatureFunctionForTransform = __rt.createSignatureFunctionForTransform;`,
+  `export const registerExportsForReactRefresh = __rt.registerExportsForReactRefresh;`,
+  `export const validateRefreshBoundaryAndEnqueueUpdate = __rt.validateRefreshBoundaryAndEnqueueUpdate;`,
+  `export const __hmr_import = __rt.__hmr_import;`,
+  `export default __rt.default || __rt;`,
+].join('\n');
+
 function getBasePath(base: string) {
   if (!base) return '/';
   if (base.startsWith('http://') || base.startsWith('https://')) {
@@ -134,6 +154,18 @@ export default function pluginDevRemoteHmr(options: NormalizedModuleFederationOp
       if (isRemote) {
         const endpointPath = getRemoteHmrPath(server.config.base);
         const wsUrl = getRemoteHmrWsUrl(server);
+
+        // Intercept /@react-refresh to serve a proxy that delegates to the
+        // host's RefreshRuntime, unifying the component registry across
+        // federation boundaries for React Fast Refresh support.
+        server.middlewares.use((req, res, next) => {
+          const url = req.url?.replace(/\?.*$/, '');
+          if (url !== '/@react-refresh') return next();
+
+          res.setHeader('Content-Type', 'application/javascript; charset=utf-8');
+          res.setHeader('Access-Control-Allow-Origin', '*');
+          res.end(REACT_REFRESH_PROXY_MODULE);
+        });
 
         server.middlewares.use((req, res, next) => {
           if (req.url?.replace(/\?.*/, '') !== endpointPath) {
